Extract memory snapshot CLI handlers
